Add 'unused' status string to fields that could be unused in certain packet types.
Some packet types don't use a certain field, or the validity of the field is
dependent on the context in which the packet was sent (ie: is it a request, or a
reply). Always reporting 'invalid' could cause confusion, as end-users might
consider this to be indicative of an error.