Only dissect 'valid fields'. Fix #6.
This changes the dissector methods to only dissect byte sequences that
are part of the 'valid fields' in packets such as JointFeedback(Ex) and
JointTrajPtFull(Ex). If a field is not in the set of valid fields, its
corresponding byte sequence will not be dissected (and thus not shown
as part of the dissection tree).

This is user configurable in the Protocol Preferences dialogue.