Facilitate dissection of captures with legacy Motoman msgs.
Those 'legacy msgs' were assigned 0x10 and 0x11 as message identifiers
during development and beta-testing of the multi-group version of the
motoman_driver package. Upon release, the Joint Traj. Pt. Full Ex. and the
Joint Feedback Ex. msgs were relocated to IDs 2016 and 2017 and the
dissector was updated accordingly.

All captures from before the relocation contain the old IDs, and this
change allows those captures to still be dissected by registering both the
old and the new ID assignments for the affected messages.

Note: this can only be done as long as IDs 0x10 and 0x11 are not assigned
to other messages.