Wait for enough bytes before reading prefix. Fix #21.
In very heavily fragmented captures it can take multiple TCP frames for
even the 'pkt_len' field to be deserialisable. This change makes the
dissector wait for at least 'sizeof(pkt_len)' bytes before continuing with
the dissection.

This is only a partial fix for #21, as the implemented approach assumes
that there are no partial messages in the TCP stream (ie: the capture was
started before any traffic was sent on any of the registered Simple
Message ports). If this is not the case, a desync will occur and the
dissection is incorrect.